bruh what

+++ Mouse chaser idk/src/Main/Main.java
@@ -25,7 +25,7 @@ public class Main extends JFrame{
 	
 	private ObjectHandler h;
 	
-	public static double score = 0;
+	public static int score = 0;
 	
 	public Main(String title) {
 		h = new ObjectHandler();
